refactor(types): move value clusters to types/tools and types/wire leaves
