+++ src/types.ts
@@ -1,4 +1,6 @@
 import type { KiroOAuthMetadata } from "./oauth/types";
+import type { OcxTool, OcxToolChoice } from "./types/tools";
+import type { UpstreamHttpVersion, ReasoningSummaryDelivery, CodexAccountMode } from "./types/wire";
 
 /** Exact provider/credential namespace for process-local reasoning replay. */
 export interface OcxReasoningReplayIdentity {
@@ -214,137 +216,17 @@ export interface OcxProviderOpaqueToolCallMetadata {
 
 export type OcxAssistantContentPart = OcxTextContent | OcxThinkingContent | OcxToolCall;
 
-export interface OcxTool {
-  name: string;
-  description: string;
-  parameters: Record<string, unknown>;
-  strict?: boolean;
-  /** MCP namespace (e.g. "mcp__context7") for tools flattened out of a Responses "namespace" tool. */
-  namespace?: string;
-  /** Freeform/custom tool (e.g. apply_patch): the model's call must be relayed as a custom_tool_call. */
-  freeform?: boolean;
-  /** Client-executed tool discovery (tool_search): the model's call must be relayed as a tool_search_call. */
-  toolSearch?: boolean;
-  /** Tool definition restored from a prior tool_search output; transports may prioritize it when catalogs are bounded. */
-  loadedFromToolSearch?: boolean;
-  /** Cursor-only synthetic exact-match edit tool; never inferred from the wire name. */
-  cursorStructuredEdit?: true;
-  /** Synthetic web_search tool: the model's call is executed by the gpt-5.4-mini sidecar, not relayed to Codex. */
-  webSearch?: boolean;
-  /** Synthetic image_gen tool: the model's call is executed by the xAI image bridge sidecar, not relayed to Codex. */
-  imageGeneration?: boolean;
-  /** Synthetic video_gen tool: executed by the xAI video bridge sidecar. */
-  videoGeneration?: boolean;
-}
-
-/**
- * Wire name a chat model sees for a tool. Namespaced (MCP) tools are flattened to
- * "<namespace>__<name>" so they survive the chat-completions function-tool format;
- * the proxy maps this back to {namespace, name} on the return trip (Codex routes MCP
- * calls by an explicit `namespace` field, not by parsing the name).
- */
-export function namespacedToolName(namespace: string | undefined, name: string): string {
-  return namespace ? `${namespace}__${name}` : name;
-}
-
-export function toolChoiceAliases(tool: Pick<OcxTool, "namespace" | "name">): string[] {
-  const wireName = namespacedToolName(tool.namespace, tool.name);
-  return tool.namespace ? [wireName, `${tool.namespace}.${tool.name}`] : [wireName];
-}
-
-function sameToolIdentity(
-  left: Pick<OcxTool, "namespace" | "name">,
-  right: Pick<OcxTool, "namespace" | "name">,
-): boolean {
-  return left.namespace === right.namespace && left.name === right.name;
-}
-
-/**
- * All tools that could be selected by one client-facing name. Bare logical names are included
- * here because they are a compatibility selector for namespaced tools, while wire and dotted
- * aliases come from `toolChoiceAliases`. A selector with more than one candidate is invalid.
- */
-export function toolChoiceCandidates(
-  tools: readonly Pick<OcxTool, "namespace" | "name">[] | undefined,
-  name: string,
-): Pick<OcxTool, "namespace" | "name">[] {
-  if (!tools) return [];
-  const candidates: Pick<OcxTool, "namespace" | "name">[] = [];
-  for (const tool of tools) {
-    if (tool.name !== name && !toolChoiceAliases(tool).includes(name)) continue;
-    if (!candidates.some(candidate => sameToolIdentity(candidate, tool))) candidates.push(tool);
-  }
-  return candidates;
-}
-
-/**
- * Newer Codex clients can select a tool nested in a namespace by its bare name. Resolve that
- * shorthand only when the request contains one tool with the logical name, so an ambiguous name
- * cannot authorize a tool from an unintended namespace.
- */
-export function toolAllowedByChoice(
-  tool: Pick<OcxTool, "namespace" | "name">,
-  allowedTools: ReadonlySet<string>,
-  tools?: readonly Pick<OcxTool, "namespace" | "name">[],
-): boolean {
-  if (!tools) return toolChoiceAliases(tool).some(name => allowedTools.has(name));
-  for (const name of [...toolChoiceAliases(tool), tool.name]) {
-    if (!allowedTools.has(name)) continue;
-    const candidates = toolChoiceCandidates(tools, name);
-    if (candidates.length === 1 && sameToolIdentity(candidates[0], tool)) return true;
-  }
-  return false;
-}
-
-export function resolveToolChoiceWireName(tools: readonly Pick<OcxTool, "namespace" | "name">[] | undefined, name: string): string {
-  const candidates = toolChoiceCandidates(tools, name);
-  if (candidates.length === 1) {
-    const match = candidates[0];
-    return namespacedToolName(match.namespace, match.name);
-  }
-  // Keep unknown/ambiguous names unchanged for callers that only serialize a selector. The
-  // catalog-aware predicate rejects them, and parseRequest rejects ambiguous request selectors.
-  return name;
-}
-
-/**
- * Whether `modelId` is in a per-provider classification list (e.g. `noVisionModels`). Matches the full
- * id, OR — for Ollama-style ids — the family before the ":size" tag, so a `gpt-oss` entry covers
- * `gpt-oss:120b`/`gpt-oss:20b`. Colon-less ids (e.g. `grok-build-0.1`) still match exactly only.
- */
-export function modelInList(list: string[] | undefined, modelId: string): boolean {
-  if (!list || list.length === 0) return false;
-  if (list.includes(modelId)) return true;
-  const colon = modelId.indexOf(":");
-  return colon > 0 && list.includes(modelId.slice(0, colon));
-}
-
-export type OcxToolChoice =
-  | "auto"
-  | "none"
-  | "required"
-  | { name: string }
-  | { allowedTools: string[]; mode: "auto" | "required" };
-
-export function isAllowedToolChoice(value: OcxToolChoice | undefined): value is { allowedTools: string[]; mode: "auto" | "required" } {
-  return typeof value === "object" && value !== null && "allowedTools" in value;
-}
-
-/** Compile the request's tool-choice policy into a reusable advertisement/restoration predicate. */
-export function toolChoiceToolPredicate(
-  choice: OcxToolChoice | undefined,
-  tools?: readonly Pick<OcxTool, "namespace" | "name">[],
-): (tool: Pick<OcxTool, "namespace" | "name">) => boolean {
-  if (!choice || choice === "auto" || choice === "required") return () => true;
-  if (choice === "none") return () => false;
-  if (isAllowedToolChoice(choice)) {
-    const allowed = new Set(choice.allowedTools);
-    return tool => toolAllowedByChoice(tool, allowed, tools);
-  }
-  if (!tools) return tool => toolChoiceAliases(tool).includes(choice.name);
-  const candidates = toolChoiceCandidates(tools, choice.name);
-  return tool => candidates.length === 1 && sameToolIdentity(candidates[0], tool);
-}
+export type { OcxTool, OcxToolChoice } from "./types/tools";
+export {
+  namespacedToolName,
+  toolChoiceAliases,
+  toolChoiceCandidates,
+  toolAllowedByChoice,
+  resolveToolChoiceWireName,
+  modelInList,
+  isAllowedToolChoice,
+  toolChoiceToolPredicate,
+} from "./types/tools";
 
 export interface OcxRequestOptions {
   maxOutputTokens?: number;
@@ -1825,86 +1707,16 @@ export interface OcxProviderConfig {
   nativeLocalExec?: "off" | "codex-sandbox" | "on";
 }
 
-/**
- * Accepted values for the per-provider upstream HTTP-version pin (#1668). Shared by the
- * zod load schema, the management write boundary (POST/PATCH), and the fetch runtime, so
- * a value that one boundary accepts can never be rejected by another.
- */
-export const UPSTREAM_HTTP_VERSION_VALUES = [
-  "auto",
-  "http1.1",
-  "h1",
-  "http2",
-  "h2",
-] as const;
-
-export type UpstreamHttpVersion = (typeof UPSTREAM_HTTP_VERSION_VALUES)[number];
-
-export const REASONING_SUMMARY_DELIVERY_VALUES = [
-  "sequential",
-  "sequential_cutoff",
-  "concurrent",
-  "concurrent_cutoff",
-] as const;
-
-export type ReasoningSummaryDelivery = typeof REASONING_SUMMARY_DELIVERY_VALUES[number];
-
-/** Trusted runtime ownership for Codex-account credentials. Never persisted per provider. */
-export type CodexAccountMode = "direct" | "pool";
-
-export const OPENAI_PROVIDER_TIER_VERSION = 2 as const;
-
-/**
- * Wires that a per-model `modelAdapters` override may select.
- *
- * Deliberately narrow: provider-specific adapters (cursor, kiro, google, ...) carry
- * their own credential and base-URL semantics, so exposing them here would widen the
- * auth boundary rather than pick a wire. Widening this set needs a per-adapter
- * credential threat model first (#404).
- */
-export const MODEL_ADAPTER_OVERRIDE_ALLOWED: ReadonlySet<string> = new Set([
-  "openai-chat",
-  "openai-responses",
-]);
-
-/**
- * Providers whose listed model ids must be driven over the Anthropic wire even when
- * the provider's configured adapter says otherwise — the upstream only speaks
- * Anthropic for these models.
- */
-const ANTHROPIC_WIRE_MODELS: Record<string, ReadonlySet<string>> = {
-  "opencode-go": new Set(["minimax-m2.5", "minimax-m2.7", "minimax-m3"]),
-};
-
-function anthropicWireModelsForProvider(providerName: string): ReadonlySet<string> | undefined {
-  return Object.hasOwn(ANTHROPIC_WIRE_MODELS, providerName)
-    ? ANTHROPIC_WIRE_MODELS[providerName]
-    : undefined;
-}
-
-/** Detached provider-local hard-pin table for pure wire-policy resolution. */
-export function captureWireAdapterHardPins(providerName: string): Readonly<Record<string, string>> {
-  const models = anthropicWireModelsForProvider(providerName);
-  if (!models) return Object.freeze({});
-  return Object.freeze(Object.fromEntries([...models].map(modelId => [modelId, "anthropic"])));
-}
-
-/**
- * True when the upstream speaks exactly one wire for this model, so a configured
- * override must not apply.
- *
- * Deliberately independent of the provider's current adapter: the wire resolver runs
- * more than once per request, and a check phrased as "pin differs from the current
- * adapter" would pass on the first pass and then let the override win on the second.
- */
-export function isWirePinnedModel(providerName: string, modelId: string): boolean {
-  return anthropicWireModelsForProvider(providerName)?.has(modelId) ?? false;
-}
-
-/** The wire a pinned model must use, or undefined when the model is not pinned. */
-export function pinnedWireAdapter(providerName: string, modelId: string): string | undefined {
-  return isWirePinnedModel(providerName, modelId) ? "anthropic" : undefined;
-}
+export type { UpstreamHttpVersion, ReasoningSummaryDelivery, CodexAccountMode } from "./types/wire";
+export {
+  UPSTREAM_HTTP_VERSION_VALUES,
+  REASONING_SUMMARY_DELIVERY_VALUES,
+  OPENAI_PROVIDER_TIER_VERSION,
+  MODEL_ADAPTER_OVERRIDE_ALLOWED,
+  captureWireAdapterHardPins,
+  isWirePinnedModel,
+  pinnedWireAdapter,
+} from "./types/wire";
 
 export interface CodexAccount {
   id: string;
